fix(appointments): TAMOC-396: Fix error when attempting to admit a patient via an outpatient appointment
